requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
django-cors-headers==4.3.1
celery==5.3.4
python-dotenv==1.0.0
Pillow==10.1.0
django-extensions==3.2.3
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7